some type changes and casts to stop the compiler warnings for pv_string

+++ src/pv2/pv_string.h
@@ -3,13 +3,15 @@
 
 #include "pv.h"
 
+#include <stdint.h>
+
 extern pv_kind string_kind;
 
 void pv_string_install();
 
 pv pv_string(const char*);
-int pv_string_length(pv);
-unsigned long pv_string_hash(pv);
+uint32_t pv_string_length(pv);
+uint32_t pv_string_hash(pv);
 pv pv_string_concat(pv, pv);
 
 #endif
